% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_names_data}
\alias{company_names_data}
\title{Company names data}
\format{An unnamed character vector.}
\source{
\url{https://www.sec.gov/help/foiadocsbdfoiahtm.html}
}
\usage{
company_names_data
}
\description{
List of registered Broker-Dealers provided by the SEC to the general public.
}
\details{
A (now outdated) extract was taken as sample input/testing data.
}
\keyword{datasets}
