% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyTFIDF.R
\name{EasyTFIDF}
\alias{EasyTFIDF}
\title{An easy function for processing TFIDF}
\usage{
EasyTFIDF(char_vector, replace_words = c(`\\t` = " "))
}
\arguments{
\item{char_vector}{A character vector of documents. To be passed as a VectorSource (tm package). The values may be duplicated but the names may not.}

\item{replace_words}{A named character vector. The element names will be replaced with the elements.}
}
\description{
This function returns an EasyTFIDF list containing handy functions
}
\examples{
super_heroes <- c('The Flash', 'The HULK', 'she-hulk', 'ant-man', 'iron-man', 'bat-man', 'super-man', 'the green arrow')
names(super_heroes) <- super_heroes
x <- EasyTFIDF(
    super_heroes, replace_words = c('-' = ' ')
)
x$doc_keys[1:3]  # look up by index
x$doc_keys['The HULK']  # see how names have been transformed
x$lookup(2:3)  # breakdown docs by dimension scores
x$CosineSimVector(2, return_sorted = T, top = 3)  # top 3 for she-hulk
x$CosineSimMatrix(2)  # all results
}
\keyword{easy}
\keyword{fuzzy}
\keyword{matching,}
\keyword{nlp,}
\keyword{tfidf,}
\keyword{tm,}
