% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_names_data}
\alias{company_names_data}
\title{A public dataset of company names}
\format{An unnamed character vector.}
\source{
\url{https://www.sec.gov/help/foiadocsbdfoiahtm.html}
}
\usage{
company_names_data
}
\description{
List of registered Broker-Dealers provided by the SEC to the general public.
}
\details{
The sample dataset provided in the eztfidf package is a good example of where the tfidf adjustment is a critical improvement over simple string distances.
Overused domain-specific language (such as 'securities' or 'associates') can be effectively mitigated with this approach.
}
\keyword{datasets}
